fix code_check

+++ frameworks/js/camera_napi/src/mode/profession_session_napi.cpp
@@ -1044,8 +1044,8 @@ napi_value ProfessionSessionNapi::ProcessingPhysicalApertures(napi_env env,
     napi_value result = nullptr;
     napi_create_array(env, &result);
     size_t zoomRangeSize = 2;
-    int zoomMinIndex = 0;
-    int zoomMaxIndex = 1;
+    size_t zoomMinIndex = 0;
+    size_t zoomMaxIndex = 1;
     for (size_t i = 0; i < physicalApertures.size(); i++) {
         if (physicalApertures[i].size() <= zoomRangeSize) {
             continue;
